update logic filter list pawn update api

+++ lib/data/response/pawn/pawn_list/pawn_list_response.dart
@@ -54,7 +54,7 @@ class ContentResponse {
   @JsonKey(name: 'availableLoanPackage')
   int? availableLoanPackage;
   @JsonKey(name: 'interest')
-  int? interest;
+  double? interest;
   @JsonKey(name: 'reputation')
   int? reputation;
   @JsonKey(name: 'avatar')
